VisionDiff_SemanticAIAuditor/requirements.txt
google-genai>=1.0.0
playwright>=1.40.0
rich>=13.0.0
tenacity>=8.2.0
structlog>=24.1.0
python-dotenv>=1.0.0
pydantic>=2.0.0
Pillow>=10.0.0
